[NO-TICKET] Bump required Ruby version to 2.5
**What does this PR do?**

This PR bumps the required ruby version for the gem to Ruby 2.5.0.

**Motivation:**

We're not actively targetting any older release than that (or testing
with), so might as well clarify that in our gemspec.

**Additional Notes:**

N/A

**How to test the change?**

Validate CI is still green :)

Validate that the required Ruby version shows up on rubygems.org
once we publish the update.