[PROF-11680] Add Ruby 3.5.0-preview1 headers and bump gem version to 3.4.1
**What does this PR do?**

This PR imports the Ruby 3.5.0-preview1 headers into the gem.

**Motivation:**

These headers are needed to get the Datadog Ruby profiler to work
correctly on Ruby 3.5.0-preview1.

(I tried without these updated headers and it broke since it will try
to fall back to the 3.4 headers).

**Additional Notes:**

This PR is very similar to
https://github.com/DataDog/datadog-ruby_core_source/pull/14 .

**How to test the change?**

I've tested this PR by running the ruby profiler test suite on
both Ruby 3.4.1 and 3.5.0-preview1 with these changes, and validating
it was green.

At some point we'll have CI in dd-trace-rb for Ruby 3.5, but currently
this testing was done locally.