[PROF-13115] Bump gem version to 3.4.2
**What does this PR do?**

This PR bumps the gem version to 3.4.2 in preparation for releasing
Ruby 4.0.0-preview2 support (added in #18).

It also updates the `test-compilation.yml` checks to run with Ruby 4.

**Motivation:**

Release a new version with support for 4.0.0-preview2.

**Additional Notes:**

**How to test the change?**

The CI check validates this change + I've tested it locally by
running `bundle exec rake package`, installing the result, and then
confirming the profiler successfully builds with it
(tested Ruby 3.4 and 4.0 manually).