[PROF-13115] Revert "Temporarily skip broken Ruby 4 test to unblock gem release"
**What does this PR do?**

This reverts commit 551b5ae11310c5bda27205cd3785c04542abe222 from #20.

See #20 for details. In that PR I mentioned:

So to break the loop, let's:

1. omit this test temporarily: ✅ done -> this was #20
2. then we can release this gem: ✅ done ->
   https://github.com/DataDog/datadog-ruby_core_source/actions/runs/20059401843,
   [version 3.4.2 is on rubygems.org](https://rubygems.org/gems/datadog-ruby_core_source/versions/3.4.2)
3. then we can undo the dd-trace-rb pinning: ✅ done ->
   https://github.com/DataDog/dd-trace-rb/pull/5122
4. then we can re-enable the test

This PR is step 4 and gets us back to where we were before the
workaround.

**Motivation:**

Remove workaround now that it's not needed anymore!

**Additional Notes:**

N/A

**How to test the change?**

Validate that Ruby 4.0 + "corruption" test is running in CI and green.