Add Ruby 4.0.0-preview3 headers
The 4.0.0-preview2 headers are not compatible with -preview3 --
in particular there's a new field in the thread structure that changes
the offset of everything that comes after it, and thus makes the
profiler crash on preview3 when compiled with preview2 headers.

I followed the usual instructions from the `README.md` to extract
the headers.